refactor : step0 피드백 반영
